requirements.txt
bcrypt
pycryptodome
requests
asyncio
flask
flask[async]
Flask-SQLAlchemy